% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tutorial_path}
\alias{get_tutorial_path}
\title{Get current tutorial file path}
\usage{
get_tutorial_path(absolute = TRUE)
}
\arguments{
\item{absolute}{}
}
\value{
NULL or the absolute path of the tutorial
}
\description{
Get current tutorial file path
}
\keyword{internal}
