% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_vlab}
\alias{initialize_vlab}
\title{Initialize the virtual lab package
It is called once in zzz.R when the package is being use}
\usage{
initialize_vlab()
}
\description{
Initialize the virtual lab package
It is called once in zzz.R when the package is being use
}
\keyword{internal}
