% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{json_filesystem_storage}
\alias{json_filesystem_storage}
\title{Filesystem-based JSON storage for tutor state data}
\usage{
json_filesystem_storage()
}
\arguments{
\item{dir}{Directory to store state data within}
}
\value{
Storage handler suitable for \code{options(tutorial.storage = ...)}
}
\description{
Tutorial state storage handler that uses the filesystem
as a backing store. The JSON file will be created in the dir,
for saving the tutorial data.
}
