% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{learnr_elements}
\alias{learnr_elements}
\alias{review_table}
\alias{note}
\alias{print_button}
\alias{print_button_logic}
\title{Learnr addon elements}
\usage{
review_table()

note(text)

print_button()

print_button_logic()
}
\description{
The following are helper element for learnr tutorials, that could extend the functionality of learnR

Note that when including these functions in a learnr Rmd document it is necessary that
the logic functions, `*_logic()`, be included in an R chunk where `context="server"` as
they interact with the underlying Shiny functionality. Conversely, any of the ui functions,
`*_ui()`, must *not* be included in an R chunk with a `context`. Both types of functions
have been written to provide useful feedback if they detect they are in the wrong R chunk
type.

Show a question review table

Note element that show a note block

Button that used to print the lab, must used with print_button_logic()

On click listener for print_button()
}
