% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_ack.R
\name{question_text_no_ack}
\alias{question_text_no_ack}
\title{Create question text with no ACK}
\usage{
question_text_no_ack(...)
}
\arguments{
\item{...}{}
}
\description{
Create question text with no ACK
}
