% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_option}
\alias{set_option}
\title{Set options and check option
If env_name is set,the function will check if the env var or option is defined
if env_name is not set, it will just check is the option is defined
this set_option util only required for}
\usage{
set_option(option_name = NULL, env_name = NULL, default = NULL)
}
\description{
Set options and check option
If env_name is set,the function will check if the env var or option is defined
if env_name is not set, it will just check is the option is defined
this set_option util only required for
}
\keyword{internal}
